move yjs into package; add web-yjs
