+++ compose/dev/compose.yml
@@ -82,7 +82,7 @@ services:
     environment:
       KERNEL_TTL: "43200000"
       LOOP_INTERVAL: "60000"
-      YJS_WS_URL: "ws://yjs:4233/socket"
+      YJS_WS_URL: "ws://web-yjs:4233/socket"
       JWT_SECRET: ${JWT_SECRET}
 
   spawner-docker:
@@ -114,12 +114,12 @@ services:
       WS_RUNTIME_IMAGE: "lihebi/codepod-runtime:0.4.13-alpha.49"
       # spawner need to add routes to proxy server
       # PROXY_API_URL: "http://proxy:4011/graphql"
-      YJS_WS_URL: "ws://yjs:4233/socket"
+      YJS_WS_URL: "ws://web-yjs:4233/socket"
       JWT_SECRET: ${JWT_SECRET}
 
-  yjs:
+  web-yjs:
     image: node:18
-    working_dir: /codepod/apps/yjs
+    working_dir: /codepod/apps/web-yjs
     ports:
       - 4233:4233
     volumes:
