+++ packages/ui/src/lib/store/yjsSlice.ts
@@ -5,7 +5,7 @@ import { produce } from "immer";
 
 import { Doc, Transaction } from "yjs";
 import * as Y from "yjs";
-import { WebsocketProvider } from "../utils/y-websocket";
+import { WebsocketProvider } from "@codepod/yjs/src/y-websocket";
 import { addAwarenessStyle } from "../utils/utils";
 import { MyState, Pod } from ".";
