+++ packages/yjs/package.json
@@ -1,7 +1,8 @@
 {
-  "name": "yjs",
+  "name": "@codepod/yjs",
   "version": "1.0.0",
   "license": "MIT",
+  "main": "src/index.ts",
   "scripts": {
     "build": "tsc",
     "start": "node build/run.js",
@@ -10,7 +11,6 @@
   },
   "dependencies": {
     "@apollo/client": "^3.7.1",
-    "@codepod/prisma": "workspace:*",
     "@kubernetes/client-node": "^0.17.1",
     "apollo-server": "^3.5.0",
     "apollo-server-core": "^3.10.3",
