+++ packages/yjs/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./yjs-setupWS";
+export * from "./yjs-utils";
