+++ packages/yjs/src/rich-schema.ts
